requirements.txt
#from qiuye
accelerate==0.33.0
transformers==4.44.0
diffusers[torch]==0.25.0
ftfy==6.1.1
# albumentations==1.3.0
opencv-python==4.8.1.78
einops==0.7.0
pytorch-lightning==1.9.0
bitsandbytes==0.46.0
lion-pytorch==0.1.2
schedulefree==1.4
pytorch-optimizer==3.7.0
prodigy-plus-schedule-free==1.9.0
prodigyopt==1.1.2
tensorboard==2.10.1
safetensors==0.4.4
prodigy-plus-schedule-free
# gradio==3.16.2
altair==4.2.2
easygui==0.98.3
toml==0.10.2
voluptuous==0.13.1
huggingface-hub==0.24.5
# for Image utils
imagesize==1.4.1
# for T5XXL tokenizer (SD3/FLUX)
sentencepiece==0.2.0
# for ui
rich==13.7.0
pandas
scipy
requests
pillow
numpy==1.26.4
# <=2.0.0
gradio==3.44.2
fastapi==0.95.1
uvicorn==0.22.0
wandb==0.16.2
httpx==0.24.1
# extra
open-clip-torch==2.20.0
lycoris-lora==2.1.0.post3
dadaptation==3.1

#sd15-tagger
tensorflow
onnx
onnxruntime-gpu

#flask-db
